% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_sample.R
\name{adapt_color_sample}
\alias{adapt_color_sample}
\title{Adapt samples specified in COLOR_SAMPLE}
\usage{
adapt_color_sample(COLOR_SAMPLE, SAMPLE)
}
\arguments{
\item{COLOR_SAMPLE}{list with groups and names of sample to color in the final plot}

\item{SAMPLE}{samples to plot (default is all samples in VCF file)}
}
\value{
A 2-columns dataframe with sample-group correspondence
}
\description{
This function iterates through COLOR_SAMPLE and checks it
It the returns a little DF of 2 columns with each specified sample and corresponding group
}
