% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_analysis.R
\name{chr_analysis}
\alias{chr_analysis}
\title{Chromosome analysis of passed VCF file}
\usage{
chr_analysis(
  VCF_DATA,
  METHOD = "RAW",
  TYPE = FALSE,
  SAMPLE_DB = FALSE,
  SHAPE = FALSE,
  THRESHOLD = FALSE,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX")
)
}
\arguments{
\item{VCF_DATA}{loaded VCF file (read_vcf() output)}

\item{METHOD}{how to count (and plot) variants for each chromosome ("RAW"=raw number, "LEN"=n/chr-length(ratio), "COD"=n/chr-coding-length(ratio), "LENCOD"=n/chr-length(ratio)*chr-coding-length(ratio))}

\item{TYPE}{(optional) divide SNP and INDELs}

\item{SAMPLE_DB}{(optional) 2-column dataframe with sample-group to color plot accordingly}

\item{SHAPE}{(optional) different shape for each sample - to implement}

\item{THRESHOLD}{(optional) a threshold line to use as Y-axis - to implement}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
gene-wise sumamries and plots
}
\description{
This function takes in the loaded VCF (read_vcf()) and operates calculations for the chromosome analysis
}
