% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_order.R
\name{chr_order}
\alias{chr_order}
\title{return a vector of chromosome names in proper genomic order}
\usage{
chr_order(CHR_VECTOR)
}
\arguments{
\item{CHR_VECTOR}{vector with chromosme names}
}
\value{
ordered vector of CHR_VECTOR
}
\description{
This function takes in a vector of chromosome names and return them in proper order
}
