% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrAnalysis.R
\name{chrAnalysis}
\alias{chrAnalysis}
\title{Make a chromosome-wise analysis}
\usage{
chrAnalysis(
  VCF_FILE,
  METHOD = "RAW",
  FASTA_FILE = FALSE,
  ASSEMBLY = "hg38",
  CENTILE = 0.9,
  SHAPE = FALSE,
  SAMPLE = "ALL",
  COLOR_SAMPLE = FALSE,
  THRESHOLD = FALSE,
  CHR_NAMES = FALSE,
  CHR_Y = FALSE
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{METHOD}{how to count (and plot) variants for each chromosome ("RAW"=raw number, "LEN"=n/chr-length(ratio), "COD"=n/chr-coding-length(ratio), "LENCOD"=n/chr-length(ratio)*chr-coding-length(ratio))}

\item{FASTA_FILE}{(optional) Path to the input FASTA file}

\item{ASSEMBLY}{(optional) which assembly your VCF is (hg38/hg19)}

\item{SHAPE}{(optional) different shape for each sample}

\item{SAMPLE}{(optional) samples to plot (default is all samples in VCF file)}

\item{COLOR_SAMPLE}{(optional) list with groups and names of sample to color in the final plot}

\item{THRESHOLD}{(optional) a threshold line to use as Y-axis}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}

\item{CHR_Y}{(optional) extend the analysis to chrY}
}
\value{
chromosome-wise summary tables and plots
}
\description{
This function generates the chromosome-wise plot starting from the VCF input file
}
