% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVCFplot.R
\name{createVCFplot}
\alias{createVCFplot}
\title{Make a Manhattan plot from a VCF file}
\usage{
createVCFplot(
  VCF_FILE,
  FASTA_FILE = FALSE,
  ASSEMBLY = "hg38",
  VAR_FLAG = "POS",
  SHAPE = FALSE,
  GENE = FALSE,
  EXON = FALSE,
  SAMPLE = "ALL",
  COLOR_SAMPLE = FALSE,
  XLIM = FALSE,
  THRESHOLD = FALSE,
  ORDERED = FALSE,
  CHR_NAMES = FALSE,
  CHR_Y = FALSE,
  SPACELINE = FALSE,
  VERBOSE = TRUE
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{FASTA_FILE}{Path to the input FASTA file}

\item{ASSEMBLY}{(optional) which assembly your VCF is (hg38/hg19)}

\item{VAR_FLAG}{(optional) the VCF variable to use as Y-axis for the variants, default is just position}

\item{SHAPE}{(optional) different shape for each sample}

\item{GENE}{(optional) genes to focus the plot on}

\item{EXON}{(optional) plot exons}

\item{SAMPLE}{(optional) samples to plot (default is all samples in VCF file)}

\item{COLOR_SAMPLE}{(optional) list with groups and names of sample to color in the final plot}

\item{XLIM}{(optional) limits for Y-axis}

\item{THRESHOLD}{(optional) a threshold line to use as Y-axis}

\item{ORDERED}{(optional) if user want to have ordered variant height in the plot or random (if value NOT specified)}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}

\item{CHR_Y}{(optional) extend the analysis to chrY}

\item{SPACELINE}{(optional) show line between chromosomes}

\item{VERBOSE}{(optional) if you want steps printed to stdout}
}
\value{
the plot
}
\description{
This function generates the plot starting from the VCF input file
}
