% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnalysis.R
\name{geneAnalysis}
\alias{geneAnalysis}
\title{Make a gene-wise analysis}
\usage{
geneAnalysis(
  VCF_FILE,
  FASTA_FILE = FALSE,
  ASSEMBLY = "hg38",
  CENTILE = 0.9,
  SHAPE = FALSE,
  SAMPLE = "ALL",
  COLOR_SAMPLE = FALSE,
  THRESHOLD = FALSE,
  CHR_NAMES = FALSE,
  CHR_Y = FALSE
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{FASTA_FILE}{Path to the input FASTA file}

\item{ASSEMBLY}{(optional) which assembly your VCF is (hg38/hg19)}

\item{CENTILE}{percentile of genes to use in the plot (default is 0.9: only genes with a number of variants >90\%-centile will be plotted)}

\item{SHAPE}{(optional) different shape for each sample}

\item{SAMPLE}{(optional) samples to plot (default is all samples in VCF file)}

\item{COLOR_SAMPLE}{(optional) list with groups and names of sample to color in the final plot}

\item{THRESHOLD}{(optional) a threshold line to use as Y-axis}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}

\item{CHR_Y}{(optional) extend the analysis to chrY}
}
\value{
gene-wise summary tables and plots
}
\description{
This function generates the gene-wise plot starting from the VCF input file
}
