% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chr_num.R
\name{get_chr_num}
\alias{get_chr_num}
\title{Extract number of variants for each chromosome in the VCF file}
\usage{
get_chr_num(
  SAMPLE_GT_CLEAN,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
)
}
\arguments{
\item{SAMPLE_GT_CLEAN}{the modeled VCF file}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
A vector with number of variants in the modeled VCF file for each specified chrosomome
}
\description{
This function extract the number of variants for each chromosome from the modeled VCF file
}
