% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chr_plot_step.R
\name{get_chr_plot_step}
\alias{get_chr_plot_step}
\title{Create a vector with height to plot each variant}
\usage{
get_chr_plot_step(CHR_N, ORDERED = FALSE)
}
\arguments{
\item{CHR_N}{the vector with number of variants for each chrosomome (as returned by get_chr_num() function)}

\item{ORDERED}{(optional) if user want to have ordered variant height in the plot or random}
}
\value{
A vector with height to plot each variant
}
\description{
This function defines the Y axis coordinate to plot each variant, based on the total variants in the VCF file
}
