% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_null.R
\name{gt_null}
\alias{gt_null}
\title{return a vector with null GT removed}
\usage{
gt_null(GT_VECTOR, GT_NULL = c("0/0", "./.", "0|0", ".|.", "0", "."))
}
\arguments{
\item{GT_VECTOR}{vector object with GT (as extracted by VariantAnnotation::readVcf())}

\item{GT_NULL}{(optional) vector passing GT values to consider null (default: c( '0/0', './.', '0|0', '.|.', '0', '.' ))}
}
\value{
non null GT vector
}
\description{
This function takes in a GT vector object and returns it without null GTs
}
