% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_chr.R
\name{load_chr}
\alias{load_chr}
\title{Load a FASTA file and get chromosomes characteristics}
\usage{
load_chr(
  FASTA_FILE,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
)
}
\arguments{
\item{FASTA_FILE}{Path to the input FASTA file}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
A Seqinfo object with length of each chromosome
}
\description{
This function loads the FASTA file and returns a Seqinfo object with length for each chromosome
or with only specified chromosomes, if CHR_NAMES is given
}
