% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Create the plot based on modeled VCF}
\usage{
make_plot(
  MODELED_VCF,
  SEQINFO,
  VAR_Y,
  VAR_FLAG = "POS",
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
)
}
\arguments{
\item{MODELED_VCF}{the database from the VCF file (load_vcf() and model_vcf())}

\item{SEQINFO}{the object with chromosome characteristics: load_chr()}

\item{VAR_Y}{the Y-axis coordinates for the variants}

\item{VAR_FLAG}{(optional) the VCF variable to use as Y-axis for the variants}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
The VCF variant plot
}
\description{
This function defines the Y axis coordinate to plot each variant, based on the total variants in the VCF file
}
