% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_vcf.R
\name{model_vcf}
\alias{model_vcf}
\title{Modify a VCF file to get a row for each variant for each sample and a GT column}
\usage{
model_vcf(VCF_DATA)
}
\arguments{
\item{VCF_DATA}{the loaded VCF file}
}
\value{
modified dataframe
}
\description{
This function modify a loaded VCF (load_vcf) file.
}
