% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read a VCF file}
\usage{
read_vcf(
  VCF_FILE,
  SEQINFO,
  ASSEMBLY = "hg38",
  SAMPLE = "ALL",
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY")
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{SEQINFO}{the object with chromosome characteristics: load_chr()}

\item{ASSEMBLY}{(optional) assembly of the VCF file (default hg38)}

\item{SAMPLE}{(optional) samples to plot (default is all samples in VCF file)}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
A matrix of the infile
}
\description{
This function read a VCF file and return the readVcf() object
}
