% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_indel.R
\name{snp_indel}
\alias{snp_indel}
\title{Extract variant type (snp/indel)}
\usage{
snp_indel(VCF_DATA)
}
\arguments{
\item{VCF_DATA}{CollapsedVCF object (as extracted by VariantAnnotation::readVcf())}
}
\value{
non null GT vector
}
\description{
This function takes in a CollapsedVCF object and returns a vector: variant type (snp/indel)
}
