% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_analysis.R
\name{spike_analysis}
\alias{spike_analysis}
\title{Load a VCF file}
\usage{
spike_analysis(
  VCF_DATA,
  CENTILE = 0.9,
  SAMPLE_DB = FALSE,
  SHAPE = FALSE,
  THRESHOLD = FALSE,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX")
)
}
\arguments{
\item{VCF_DATA}{loaded VCF file (read_vcf() output)}

\item{CENTILE}{percentile of genes to use in the plot (default is 0.9: only genes with a number of variants >90\%-centile will be plotted)}

\item{SAMPLE_DB}{(optional) 2-column dataframe with sample-group to color plot accordingly}

\item{SHAPE}{(optional) different shape for each sample - to implement}

\item{THRESHOLD}{(optional) a threshold line to use as Y-axis - to implement}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}
}
\value{
gene-wise sumamries and plots
}
\description{
This function loads a VCF file as a matrix.
}
