% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_genes.R
\name{vcf_genes}
\alias{vcf_genes}
\title{return a GRanges object with gene symbols annotated}
\usage{
vcf_genes(VCF_DATA, COLUMN = "symbol", STRING = FALSE, SINGLE = FALSE)
}
\arguments{
\item{VCF_DATA}{GRanges object}

\item{COLUMN}{(optional) the column name of GENES38 to annotate (default: "symbol")}

\item{STRING}{(optional) get each gene values as comma separated string (in case of multiple genes)}

\item{SINGLE}{(optional) get only one genes per variant}
}
\value{
GRanges annotated object
}
\description{
This function takes in a GRanges object and returns it with genes annotated
}
