% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVCFplot.R
\name{createVCFplot}
\alias{createVCFplot}
\title{Make a Manhattan plot from a VCF file}
\usage{
createVCFplot(
  VCF_FILE,
  FASTA_FILE,
  ASSEMBLY = "hg38",
  VAR_FLAG = "POS",
  ORDERED = FALSE,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY"),
  VERBOSE = TRUE
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{FASTA_FILE}{Path to the input FASTA file}

\item{ASSEMBLY}{(optional) which assembly your VCF is (hg38/hg19)}

\item{ORDERED}{(optional) if user want to have ordered variant height in the plot or random (if value NOT specified)}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}

\item{VERBOSE}{(optional) if you want steps printed to stdout}

\item{VALUE}{(optional) the VCF variable to use as Y-axis for the variants, default is just position}
}
\value{
the plot
}
\description{
This function generates the plot starting from the VCF input file
}
