% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createVCFplot.R
\name{createVCFplot}
\alias{createVCFplot}
\title{This is the function that calls all the other functions in order to create the definitive plot}
\usage{
createVCFplot(
  VCF_FILE,
  FASTA_FILE,
  ORDERED = FALSE,
  CHR_NAMES = c("chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9",
    "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18",
    "chr19", "chr20", "chr21", "chr22", "chrX", "chrY"),
  VERBOSE = TRUE
)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}

\item{FASTA_FILE}{Path to the input FASTA file}

\item{ORDERED}{(optional) if user want to have ordered variant height in the plot or random}

\item{CHR_NAMES}{(optional) vector with chromosme names to plot}

\item{VERBOSE}{(optional) if you want steps printed to stdout}
}
\value{
the plot
}
\description{
This function generates the plot starting from the VCF input file
}
