% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_vcf.R
\name{load_vcf}
\alias{load_vcf}
\title{Load a VCF file}
\usage{
load_vcf(VCF_FILE)
}
\arguments{
\item{VCF_FILE}{Path to the input VCF file}
}
\value{
A matrix of the infile
}
\description{
This function loads a VCF file as a matrix.
}
